uriloader/exthandler/tests/mochitest/test_web_protocol_handlers.html

<!DOCTYPE html>
<html>
<head>
  <title>Test for disabled registerProtocolHandler</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/AddTask.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<iframe id="test"></iframe>
<script type="text/javascript">
add_task(async function() {
  await SpecialPowers.pushPrefEnv({
    set: [
      ["dom.registerContentHandler.enabled", true],
      ["dom.registerProtocolHandler.insecure.enabled", false],
    ]
  });

  let result = "registerProtocolHandler" in navigator;
  ok(!result, "Insecure registerProtocolHandler is undefined");
});
</script>
</body>
</html>
